+++ packages/atomic/src/components/search/result-template-components/quickview-preview-bar/quickview-preview-bar.ts
@@ -55,7 +55,7 @@ export const buildQuickviewPreviewBar = (
     return;
   }
   const bar = buildPreviewBar(documentWriter);
-  if (!highlightKeywords.highlightAll) {
+  if (highlightKeywords.highlightNone) {
     bar.remove();
     return;
   }