+++ packages/headless/src/features/query/query-actions.ts
@@ -9,7 +9,7 @@ export interface UpdateQueryActionCreatorPayload {
   q?: string;
 
   /**
-   * Whether to interpret advanced [Coveo Cloud query syntax](https://docs.coveo.com/en/1814/searching-with-coveo/search-prefixes-and-operators) in the query.
+   * Whether to interpret advanced [Coveo Cloud query syntax](https://docs.coveo.com/en/1814/) in the query.
    */
   enableQuerySyntax?: boolean;
 }