+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.7.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.6.0...@coveo/atomic@2.7.0) (2023-01-12)
+
+### Bug Fixes
+
+- **atomic:** removed console warning when pipeline or search hub aren't configured ([#2654](https://github.com/coveo/ui-kit/issues/2654)) ([1e97521](https://github.com/coveo/ui-kit/commit/1e9752175b7ab74162ff42d31c975f54615022a5))
+
+### Features
+
+- **atomic:** added query syntax support to search box ([#2647](https://github.com/coveo/ui-kit/issues/2647)) ([c4c96b8](https://github.com/coveo/ui-kit/commit/c4c96b83136dd4b9a4cd34d613ec2b0737673e1b))
+
+# [2.6.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.5.0...@coveo/atomic@2.6.0) (2023-01-11)
+
+### Features
+
+- **atomic,atomic-react:** added warning when initializing with engine and pipeline ([#2644](https://github.com/coveo/ui-kit/issues/2644)) ([d816158](https://github.com/coveo/ui-kit/commit/d816158f72fa0cde7250221cb10855351c6461eb))
+
+# [2.5.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.4.4...@coveo/atomic@2.5.0) (2023-01-11)
+
+### Features
+
+- **atomic:** add quickview hit highlight preview bar ([#2643](https://github.com/coveo/ui-kit/issues/2643)) ([cb3e212](https://github.com/coveo/ui-kit/commit/cb3e2127d4cdb7d3a40d1da45826e9dc9f0d7a9c))
+- **atomic:** implement quickview pagination/navigation/keyword hit ([#2623](https://github.com/coveo/ui-kit/issues/2623)) ([c5ec5e1](https://github.com/coveo/ui-kit/commit/c5ec5e10e37ce3a9b4a0c6ac3220a57d6242c7a8))
+
 ## [2.4.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.4.3...@coveo/atomic@2.4.4) (2023-01-09)
 
 **Note:** Version bump only for package @coveo/atomic