+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.4](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.7.3...@coveo/quantic@2.7.4) (2023-01-12)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.7.3](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.7.2...@coveo/quantic@2.7.3) (2023-01-11)
+
+**Note:** Version bump only for package @coveo/quantic
+
 ## [2.7.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.7.1...@coveo/quantic@2.7.2) (2023-01-09)
 
 **Note:** Version bump only for package @coveo/quantic